refactor #3 - extract out detail view
