+++ judgments/views/__init__.py
@@ -1,6 +1,5 @@
 import json
 import math
-import re
 from typing import Optional, Union
 
 import environ
@@ -18,63 +17,14 @@ from django.utils.translation import gettext
 from requests_toolbelt.multipart import decoder
 
 from judgments.models import SearchResult, SearchResults
-from judgments.utils import (
-    VERSION_REGEX,
-    create_s3_client,
-    delete_documents,
-    generate_docx_url,
-    generate_pdf_url,
-    get_judgment_root,
-    uri_for_s3,
-)
+from judgments.utils import create_s3_client, delete_documents
 
+from .detail import detail  # noqa
 from .edit_judgment import EditJudgmentView  # noqa
 
 env = environ.Env()
 akn_namespace = {"akn": "http://docs.oasis-open.org/legaldocml/ns/akn/3.0"}
 uk_namespace = {"uk": "https://caselaw.nationalarchives.gov.uk/akn"}
-
-
-def detail(request):
-    params = request.GET
-    judgment_uri = params.get("judgment_uri", None)
-    version_uri = params.get("version_uri", None)
-    context = {"judgment_uri": judgment_uri, "is_failure": False}
-
-    try:
-        judgment_xml = api_client.get_judgment_xml(judgment_uri, show_unpublished=True)
-        judgment_root = get_judgment_root(judgment_xml)
-        context["is_editable"] = True
-        if "failures" in judgment_uri:
-            context["is_failure"] = True
-
-        if "error" in judgment_root:
-            context["is_editable"] = False
-            judgment = judgment_xml
-            metadata_name = judgment_uri
-        else:
-            results = api_client.eval_xslt(
-                judgment_uri, version_uri, show_unpublished=True
-            )
-            metadata_name = api_client.get_judgment_name(judgment_uri)
-
-            multipart_data = decoder.MultipartDecoder.from_response(results)
-            judgment = multipart_data.parts[0].text
-        context["judgment"] = judgment
-        context["page_title"] = metadata_name
-        context["docx_url"] = generate_docx_url(uri_for_s3(judgment_uri))
-        context["pdf_url"] = generate_pdf_url(uri_for_s3(judgment_uri))
-
-        if version_uri:
-            try:
-                version = re.search(VERSION_REGEX, version_uri).group(1)
-            except AttributeError:
-                version = None
-            context["version"] = version
-    except MarklogicResourceNotFoundError as e:
-        raise Http404(f"Judgment was not found at uri {judgment_uri}, {e}")
-    template = loader.get_template("judgment/detail.html")
-    return HttpResponse(template.render({"context": context}, request))
 
 
 def detail_xml(request):
